reorganisation: rpm submodule
